chore: renamed get activity to issue activity
