+++ src/parser/content-evaluator-module.ts
@@ -2,7 +2,7 @@ import Decimal from "decimal.js";
 import { encodingForModel } from "js-tiktoken";
 import OpenAI from "openai";
 import configuration from "../configuration/config-reader";
-import { GetActivity } from "../get-activity";
+import { IssueActivity } from "../issue-activity";
 import { GitHubIssue } from "../github-types";
 import { GithubCommentScore, Module, Result } from "./processor";
 
@@ -17,7 +17,7 @@ export class ContentEvaluatorModule implements Module {
     return this.configuration.enabled;
   }
 
-  async transform(data: Readonly<GetActivity>, result: Result) {
+  async transform(data: Readonly<IssueActivity>, result: Result) {
     const promises: Promise<void>[] = [];
 
     for (const key of Object.keys(result)) {
